prepare routes to enable different test scenarios
