+++ apps/tasks-3.x/package.json
@@ -11,7 +11,9 @@
     "@babel/runtime": "^7.23.5",
     "meteor-node-stubs": "^1.2.7",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router": "6.3.0",
+    "react-router-dom": "6.28.1"
   },
   "meteor": {
     "mainModule": {
